Updated datatype for UKPRN to bigint

+++ src/AcceptanceTesting/Common/Create_TestStack_Db.sql
@@ -44,7 +44,7 @@ GO
 
 CREATE TABLE TestStack.Provider
 (
-	Ukprn int PRIMARY KEY,
+	Ukprn bigint PRIMARY KEY,
 	ProviderName varchar(50) NULL
 )
 GO
